extract TextStyleBuilder to own file
